fix: CUDA dtype mismatch in ultra_precision.py (v0.4.1)
Cloud Validation Results (GCP NVIDIA L4):
- 66/66 NVIDIA backend tests PASSED
- 1300/1300 NVIDIA benchmarks PASSED
- 2.37x speedup achieved (0.74ms vs 1.76ms)

Bug Fix:
- ultra_precision.py:675: Fixed RuntimeError when quantized values
  had different dtype than target tensor on CUDA
- Added .to(quantized_tensor.dtype) to ensure compatibility

Testing:
- Validated on GCP g2-standard-4 with NVIDIA L4 (23GB)
- PyTorch 2.7.1+cu128, CUDA 12.8
- 905 local tests passing

+++ src/kernel_pytorch/precision/ultra_precision.py
@@ -671,8 +671,8 @@ class UltraPrecisionModule(nn.Module):
                 # Apply precision-specific quantization
                 quantized_values = self._quantize_values(level_values, precision_fmt)
 
-                # Place back in tensor
-                quantized_tensor[level_mask] = quantized_values
+                # Place back in tensor (ensure dtype matches to avoid CUDA errors)
+                quantized_tensor[level_mask] = quantized_values.to(quantized_tensor.dtype)
 
         return quantized_tensor
